Moved pitch analysis functions to the Analyzer component, out of Game.
