+++ src/components/Analyzer.js
@@ -1,16 +1,55 @@
 import React, { Component } from 'react';
-// import a component if you want to return it in your render
 import Pitchfinder from 'pitchfinder';
+// import a component if you want to return it in your render
 
 class Analyzer extends Component {
+  constructor(){
+    super();
+    this.state={
+      processor: null
+    }
 
-  detectPitch = Pitchfinder.YIN();
+  }
 
-  // myAudioBuffer = this.props.userAudioFromMic.getAudioBuffer(); // assume this returns a WebAudio AudioBuffer object
-  // float32Array = this.myAudioBuffer.getChannelData(0); // get a single channel of sound
-  // pitch = this.detectPitch(float32Array); // null if pitch cannot be identified
+  componentDidMount(){
+    this.runAnalyzer();
+  }
 
-// take mic input and .connect to audioBuffer?
+  componentWillUnmount() {
+    if (this.state.processor){this.state.processor.disconnect()}; // without this, error: Cannot read property 'disconnect' of null
+  }
+
+  async runAnalyzer() {
+    const userAudioFromMic = await navigator.mediaDevices.getUserMedia({
+      audio: true,
+      video: false
+    });
+    // console.log('userAudioFromMic: ', userAudioFromMic);
+    // userAudioFromMic is now accessible in state as a MediaStream
+
+     const context = new AudioContext();
+     const source = context.createMediaStreamSource(userAudioFromMic);
+     const processor = context.createScriptProcessor(4096, 1, 1);
+     //Creates a ScriptProcessorNode for direct audio processing.
+     // Arguments: bufferSize, numberOfInputChannels, numberOfOutputChannels
+     // bufferSize: no. of units of sample-frames; values must be: 256, 512, 1024, 2048, 4096, 8192, or 16384. This value controls how frequently the onaudioprocess event handler is called and how many sample-frames need to be processed each call.
+     this.setState({processor: processor}); // the ScriptProcessorNode is now available in state as processor
+     source.connect(processor); // route the output of the source to the input of the processor; userAudioFromMic -->> ScriptProcessorNode
+     processor.connect(context.destination); // routes the output of the processor node to the destination node; required
+     // console.log('processor', processor);
+     const detectPitch = Pitchfinder.AMDF();
+     // console.log('Pitchfinder.AMDF', Pitchfinder.AMDF);
+
+
+     processor.onaudioprocess = function(audioBuffer) {
+       const pitch = detectPitch(audioBuffer.inputBuffer.getChannelData(0));
+       // get a single channel of sound from the AudioBuffer object
+       // const pitch = detectPitch(float32Array);
+       // each time the buffer is added to, the pitch is detected from the input
+       // null if pitch cannot be identified
+       console.log('pitch: ', pitch);
+      }
+  }
 
   render(){
     return(<div></div>);
@@ -19,12 +58,3 @@ class Analyzer extends Component {
 }
 
 export default Analyzer;
-
-
-
-// const Pitchfinder = require("pitchfinder");
-// const detectPitch = Pitchfinder.AMDF();
-//
-// const myAudioBuffer = getAudioBuffer(); // assume this returns a WebAudio AudioBuffer object
-// const float32Array = myAudioBuffer.getChannelData(0); // get a single channel of sound
-// const pitch = detectPitch(float32Array); // null if pitch cannot be identified
